Auth Now has functionality. Test auth controller made
